% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_iris_data.R
\docType{data}
\name{iris_data}
\alias{iris_data}
\title{iris_data}
\format{
A data frame with 150 rows and 6 variables:
\describe{
\item{ V1 }{  integer }
\item{ Sepal.Length }{  numeric }
\item{ Sepal.Width }{  numeric }
\item{ Petal.Length }{  numeric }
\item{ Petal.Width }{  numeric }
\item{ Species }{  character }
}
}
\source{
Source
}
\usage{
iris_data
}
\description{
Description.
}
\keyword{datasets}
