% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary.R
\name{my_summary}
\alias{my_summary}
\title{Title}
\usage{
my_summary(df, by_col)
}
\arguments{
\item{df}{A data frame or data table}

\item{by_col}{group by column}
}
\value{
a Data Frame
}
\description{
Description
}
\examples{
data("iris_data")
my_summary(df = iris,
           by = "Species")
}
