% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscatter.R
\name{myScatter}
\alias{myScatter}
\title{Title}
\usage{
myScatter(df, x, y, col_var = NULL)
}
\arguments{
\item{df}{A data frame}

\item{x, }{x axis value}

\item{y}{y axis value}

\item{col_var}{color variable value}
}
\value{
a list of class ggplot
}
\description{
Description
}
\examples{
myScatter(df = iris_data,
          x = Petal.Width , 
          y = Petal.Length , 
          col_var = Species)
}
