% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytoolsPackage-package.R
\docType{package}
\name{mytoolsPackage-package}
\alias{mytoolsPackage}
\alias{mytoolsPackage-package}
\title{mytoolsPackage: A Package With A Few Functions}
\description{
A package with some things.
}
\author{
\strong{Maintainer}: Moses Mburu \email{mmburu20@gmail.com}

}
\keyword{internal}
